backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg[binary]==3.3.3
python-dotenv==1.0.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==1.3.1
python-jose==3.3.0
pytest==7.4.3
supabase==2.0.1
python-multipart==0.0.6
